requirements/base.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
drf-spectacular>=0.26.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
dj-database-url>=2.1.0